+++ src/Canbus/Canbus.h
@@ -76,7 +76,7 @@ class Canbus
 
         void sendMessage(
             uint8_t priority,
-            uint8_t serviceTypeId,
+            uint16_t serviceTypeId,
             uint8_t destNodeId,
             uint8_t *payload,
             uint8_t payloadLength
